# specs/002-admin-panel-telegram-auth/contracts/auth-admin-api.openapi.yaml
openapi: 3.1.0
info:
  title: BasmatLibya — Auth + Admin API (feature 002)
  version: 0.2.0
  description: |
    Additive REST surface for feature `002-admin-panel-telegram-auth`. This document
    describes ONLY the new endpoints introduced by this feature: Telegram-based
    authentication, signed-in user history, and the owner-only admin panel.

    The v1 surface from `001-digital-footprint-analyzer` (`/api/lookups`,
    `/api/lookups/:id`, `/api/lookups/:id/rerun`, `/api/healthz`) remains
    unchanged. All schemas here are generated from `zod` schemas in
    `packages/shared`; treat this file as a snapshot — the runtime is the source
    of truth.

    Authentication: a server-issued opaque session cookie `bsl_session`
    (HttpOnly; Secure; SameSite=Lax). State-changing endpoints additionally
    require an `X-CSRF` header equal to the session's `csrfToken` (returned
    by `/api/auth/me`). The Telegram callback `/api/auth/telegram` is exempt
    from `X-CSRF` because it carries its own HMAC verification.

    Authorization roles:
      - anonymous: no cookie or revoked cookie.
      - user:     valid cookie, role='user', status='active'.
      - owner:    valid cookie, role='owner', status='active'.

    All user-facing strings on `4xx`/`5xx` errors are Arabic and pre-translated
    server-side; clients render them verbatim.
servers:
  - url: https://basmatlibya.example.com
    description: Production (Render)
  - url: http://localhost:5173
...
paths:
...
  /api/admin/ai-models:
    get:
      tags: [admin-ai-models]
      summary: List AI model entries
      operationId: listAiModels
      responses:
        '200':
          description: Entries (credentials are never returned in plaintext).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AiModelEntryDisplay' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    post:
      tags: [admin-ai-models]
      summary: Create a new AI model entry
      description: |
        On create, the server validates the credentials against the provider
        before persisting (FR-012). On validation failure the row is NOT
        persisted as `active`; the response is `400` with a designed Arabic
        message.
      operationId: createAiModel
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiModelEntryCreate' }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiModelEntryDisplay' }
        '400':
          description: Validation failed (provider rejected credentials, params out of range, etc.).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/admin/ai-models/{id}:
    patch:
      tags: [admin-ai-models]
      summary: Update an AI model entry (re-validates on save)
      operationId: updateAiModel
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiModelEntryUpdate' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiModelEntryDisplay' }
        '400':
          description: Validation failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Entry not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [admin-ai-models]
      summary: Delete an AI model entry
      description: |
        Deleting the currently active entry is rejected (`409`) unless
        another entry has been activated first or `enrichment_enabled`
        is `false`. See FR-013 and the admin edge-case in `spec.md`.
      operationId: deleteAiModel
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Deleted.
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Entry not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Cannot delete the currently active entry.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/admin/ai-models/{id}/activate:
    post:
      tags: [admin-ai-models]
      summary: Designate an AI model entry as the active model
      description: |
        Atomically clears `is_active` on every other row and sets it on
        the target. Re-validates the target's credentials before flipping
        the flag (FR-012, FR-013).
      operationId: activateAiModel
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Active model switched.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiModelEntryDisplay' }
        '400':
          description: Re-validation failed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Caller is not the owner.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Entry not found.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  parameters:
    CsrfHeader:
      in: header
      name: X-CSRF
      required: true
      schema: { type: string }
      description: Per-session double-submit token from `/api/auth/me` (R-04).
  schemas:
...
    AiModelProvider:
      type: string
      enum: [openai, anthropic, google, nvidia, openai_compatible]
    AiModelStatus:
      type: string
      enum: [active, inactive, invalid]
    AiModelGenerationParams:
      type: object
      required: [systemPrompt, temperature, maxOutputTokens, extraParams]
      properties:
        systemPrompt:     { type: string }
        temperature:      { type: number, minimum: 0, maximum: 2 }
        maxOutputTokens:  { type: integer, minimum: 16, maximum: 32000 }
        extraParams:      { type: object, additionalProperties: true }
    AiModelEntryCreate:
      type: object
      required: [provider, modelId, credential, generation]
      properties:
        provider:     { $ref: '#/components/schemas/AiModelProvider' }
        modelId:      { type: string }
        displayLabel: { type: string, nullable: true }
        credential:   { type: string, description: "Plaintext credential — accepted on the wire only over TLS, encrypted at rest with MODEL_SECRET_KEY." }
        generation:   { $ref: '#/components/schemas/AiModelGenerationParams' }
    AiModelEntryUpdate:
      type: object
      additionalProperties: false
      properties:
        modelId:      { type: string }
        displayLabel: { type: string, nullable: true }
        credential:   { type: string, description: "Optional rotation. If omitted, the existing ciphertext is preserved." }
        generation:   { $ref: '#/components/schemas/AiModelGenerationParams' }
    AiModelEntryDisplay:
      type: object
      required: [id, provider, modelId, displayLabel, credential, generation, status, isActive, validatedAt, createdAt, lastUpdatedAt]
      properties:
        id:           { type: string, format: uuid }
        provider:     { $ref: '#/components/schemas/AiModelProvider' }
        modelId:      { type: string }
        displayLabel: { type: string, nullable: true }
        credential:
          type: object
          required: [present, lastFour]
          description: "Server-side display shape — plaintext is never returned (G8)."
          properties:
            present:  { type: boolean }
            lastFour: { type: string, description: 'Last 4 plaintext chars, e.g. "•••• 1234".' }
        generation:        { $ref: '#/components/schemas/AiModelGenerationParams' }
        status:            { $ref: '#/components/schemas/AiModelStatus' }
        isActive:          { type: boolean }
        validatedAt:       { type: string, format: date-time, nullable: true }
        lastValidationError: { type: string, nullable: true }
        createdAt:         { type: string, format: date-time }
        lastUpdatedAt:     { type: string, format: date-time }
...
    ErrorResponse:
      type: object
      required: [code, messageAr]
      properties:
        code:
          type: string
          description: Stable machine code.
          enum:
            - not_authenticated
            - unauthorized
            - csrf_required
            - hmac_invalid
            - auth_date_too_old
            - bot_unavailable
            - suspended_user
            - validation_failed
            - last_owner_protected
            - dependent_entries
            - active_model_protected
            - rate_limited
            - not_found
        messageAr:
          type: string
          description: Pre-translated Arabic message; the UI renders verbatim.
        details:
          type: object
          nullable: true
          description: Optional structured details (e.g. per-field validation errors).